refactor: common native interface
We refactor the sources to establish a common native interface.
This also allows us to tidy up the coding by moving big chunks of
platform-dependent implementations into dedicated sources,
streamlining the main logic.
